requirements.txt
bcrypt
flask
pyjwt